telegram-bot/requirements.txt
# telegram-bot/requirements.txt
# Estif Bingo 24/7 - Python Dependencies
# Run: pip install -r requirements.txt

# ==================== CORE FRAMEWORKS ====================
python-telegram-bot==20.7
Flask==3.0.0
flask-socketio==5.3.4
python-socketio==5.10.0
eventlet==0.33.3

# ==================== DATABASE ====================
asyncpg==0.29.0
SQLAlchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# ==================== AUTHENTICATION & SECURITY ====================
PyJWT==2.8.0
cryptography==41.0.7
passlib==1.7.4
bcrypt==4.1.2
python-dotenv==1.0.0

# ==================== API & WEB ====================
Flask-CORS==4.0.0
requests==2.31.0
aiohttp==3.9.1

# ==================== DATA VALIDATION & SERIALIZATION ====================
pydantic==2.5.0
pydantic-settings==2.1.0
marshmallow==3.20.1
python-multipart==0.0.6

# ==================== LOGGING & MONITORING ====================
python-json-logger==2.0.7
prometheus-client==0.19.0
sentry-sdk==1.38.0

# ==================== UTILITIES ====================
python-dateutil==2.8.2
pytz==2023.3
chardet==5.2.0
emoji==2.8.0
pillow==10.1.0

# ==================== ASYNC & CONCURRENCY ====================
asyncio==3.4.3
aiofiles==23.2.1
backoff==2.2.1

# ==================== TESTING (Development only) ====================
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# ==================== DEPLOYMENT ====================
gunicorn==21.2.0
uvloop==0.19.0